requirements.txt
aiogram==2.20
aiogram-dialog==1.4.2
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.13.1
environs==9.5.0
et-xmlfile==1.1.0
frozenlist==1.3.0
idna==3.3
Jinja2==3.1.2
magic-filter==1.0.7
MarkupSafe==2.1.1
marshmallow==3.16.0
multidict==6.0.2
openpyxl==3.0.10
packaging==21.3
Pillow==9.1.1
psycopg2-binary==2.9.3
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
screen==1.0.1
sqlparse==0.4.2
typing-extensions==4.2.0
xlwt==1.3.0
yarl==1.7.2
